Incorrect argument index, there is only one argument as per session.xml
This causes an OutOfBoundsException, I don't believe this was tested. If you set your session handler to null, which we did when upgrading from Symfony 2.1 to Symfony 2.3 we got the exception. Looking at the session.xml it seems quite clear that it's the session handler that we wish to null out.

Further to that, there is a code smell in the PhpBridgeSessionStorage.php, it's constructor definition causes it to violate the liskov substitution principle. If that wasn't the case this entire problem would go away, the definition changes here could look the same and would have the arguments in the correct/same positions.